Moved util to specified folder
